Install skyline-console through yarn
At the moment it is not possible to install skyline-console from repo
through pip since static files were updated last time 2 years ago,
and they do not correspond to relevant content of the project.

With that, it is still possible to install skyline-console with pip but only
from PyPi. However, Skyline project does not cut releases frequently,
so it's worth enabling building skyline-console with yarn in case
of need for urgent bugfix rather then wait for another PyPi release.

For that we add logic that will attempt to discover if we're dealing with
PyPi release or some SHA and act accordingly to this.
We also tring to be smart on when to re-build skyline-console
by fetching actual state/available wheels on repo container.

Change-Id: Iacb703b5ee4d51990edc3dea717b3b8b57705a60