Enable SSLProxyProtocol for internal backends behind TLS
In cases when internal backends are covered with TLS
(ie
`haproxy_ssl_all_vips` is set to True), Apache will fail to ProxyPass
traffic if it needs to be encrypted. For that, SSL module should be
installed and enabled as well as SSLProxyEngine should be turned on.

We also introduce SSL protocol and cipher variables which will be used
in the future to cover connection between HAProxy and Apache in the
future, while now service the same purpose for the TLS proxy.

Closes-Bug: #2077904
Change-Id: I7e7c7053cce96d774199f37c7255e70c49e2969a
(cherry picked from commit 93b7ef9b2546d0998880877c682b8d2dffd43e4c)